\name{jkmeans-package}
\alias{jkmeans-package}
\alias{jkmeans}
\docType{package}
\title{
\packageTitle{jkmeans}
}
\description{
\packageDescription{jkmeans}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jkmeans}
\packageIndices{jkmeans}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{jkmeans}

Maintainer: \packageMaintainer{jkmeans}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
